Shutdown supervisord properly. We can't launch automatically now.
We can't relaunch supervisord for a biazzare bug/problem in this automation script which I still can't understand. The new path we added to /etc/enviornment is not reloaded. I can understand that not every process can obtain the latest, but restarting a process should be able to acquire the latest changes in /etc/environment. Even doing ``raw: printenv`` I can see the new path.